+++ src/GraphCanvas.tsx
@@ -141,8 +141,12 @@ export const GraphCanvas: FC<GraphCanvasProps & { ref?: Ref<GraphCanvasRef> }> =
       useImperativeHandle(ref, () => ({
         centerGraph: (nodeIds, opts) =>
           rendererRef.current?.centerGraph(nodeIds, opts),
+        fitNodesInView: (nodeIds, opts) =>
+          rendererRef.current?.fitNodesInView(nodeIds, opts),
         zoomIn: () => controlsRef.current?.zoomIn(),
         zoomOut: () => controlsRef.current?.zoomOut(),
+        dollyIn: distance => controlsRef.current?.dollyIn(distance),
+        dollyOut: distance => controlsRef.current?.dollyOut(distance),
         panLeft: () => controlsRef.current?.panLeft(),
         panRight: () => controlsRef.current?.panRight(),
         panDown: () => controlsRef.current?.panDown(),